+++ packages/template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @typed-assistant/template
 
+## 0.0.103
+
+### Patch Changes
+
+- Updated dependencies
+  - @typed-assistant/builder@0.0.85
+
 ## 0.0.102
 
 ### Patch Changes